requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.3.0
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
black==22.6.0
bleach==6.0.0
catboost==1.1.1
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.0.4
cloudpickle==2.2.1
contourpy==1.0.7
cycler==0.11.0
dask==2023.3.1
DAWG-Python==0.7.2
debugpy==1.5.1
decorator==5.1.1
docker-pycreds==0.4.0
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.1
huggingface-hub==0.13.1
idna==3.4
imageio==2.26.0
ipykernel==6.15.0
ipython==8.11.0
ipywidgets==8.0.4
jedi==0.18.2
joblib==1.2.0
jupyter-client==7.3.4
jupyter_core==5.2.0
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lazy_loader==0.1
lightning-utilities==0.8.0
locket==1.0.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
multidict==6.0.4
mypy-extensions==0.4.3
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
parso==0.8.3
partd==1.3.0
pathspec==0.10.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.0.1
platformdirs==3.1.0
plotly==5.13.1
prompt-toolkit==3.0.38
protobuf==4.22.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
Pygments==2.14.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.9.4
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==23.2.0
qudida==0.0.4
regex==2022.10.31
requests==2.28.2
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.16.0
setproctitle==1.3.2
setuptools==65.6.3
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
swifter==1.3.4
tenacity==8.2.2
threadpoolctl==3.1.0
tifffile==2023.2.28
tokenizers==0.13.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchmetrics==0.11.4
torchvision==0.14.1
tornado==6.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typing_extensions==4.4.0
urllib3==1.26.15
wandb==0.13.11
wcwidth==0.2.6
webencodings==0.5.1
wheel==0.38.4
widgetsnbextension==4.0.5
yarl==1.8.2
